+++ src/core/messaging/messages/replymessage.h
@@ -10,7 +10,7 @@ class ReplyMessage : public Message
 {
 public:
     ReplyMessage(const ClusterID targetClusterId,
-                 const uint32_t messageId,
+                 const uint64_t messageId,
                  const uint8_t targetSite);
     ReplyMessage();
